Refactor : 6주차 멘토님 코드리뷰 반영
