+++ src/pages/common/register/index.ts
@@ -3,4 +3,3 @@ export { default as RegisterPage } from './RegisterPage';
 export * from './data';
 export * from './components';
 export * from './types';
-export * from './api';
