+++ src/pages/common/register/store/api/index.ts
@@ -1,0 +1,1 @@
+export * from './user-register';
