+++ src/pages/common/register/store/hooks/index.ts
